+++ .php-cs-fixer.dist.php
@@ -3,6 +3,7 @@
 $finder = PhpCsFixer\Finder::create()
     ->in(__DIR__ . '/src')
     ->in(__DIR__ . '/tests')
+    ->in(__DIR__ . '/rector')
     ->name('*.php')
     ->ignoreDotFiles(true)
     ->ignoreVCS(true);